Replace generic yes/no dialog with more specific one for casting confirmation.
This is to work around the fact that we can't serialize a Runnable, and have constructors with arguments is not good for dialogs.
